back-end/mvp/requirements.txt
alabaster==0.7.12 ; python_version >= "3.10" and python_version < "4.0"
anyio==3.6.1 ; python_version >= "3.10" and python_version < "4.0"
asyncpg==0.26.0 ; python_version >= "3.10" and python_version < "4.0"
attrs==22.1.0 ; python_version >= "3.10" and python_version < "4.0"
babel==2.10.3 ; python_version >= "3.10" and python_version < "4.0"
backports-cached-property==1.0.2 ; python_version >= "3.10" and python_version < "4.0"
bcrypt==4.0.0 ; python_version >= "3.10" and python_version < "4.0"
black==22.10.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2022.9.24 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.15.1 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==2.1.1 ; python_version >= "3.10" and python_version < "4"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.5 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32" or python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
contextlib2==21.6.0 ; python_version >= "3.10" and python_version < "4.0"
crud==0.1 ; python_version >= "3.10" and python_version < "4.0"
cryptography==38.0.1 ; python_version >= "3.10" and python_version < "4.0"
dependency-injector==4.40.0 ; python_version >= "3.10" and python_version < "4.0"
docutils==0.16 ; python_version >= "3.10" and python_version < "4.0"
ecdsa==0.18.0 ; python_version >= "3.10" and python_version < "4.0"
fastapi==0.78.0 ; python_version >= "3.10" and python_version < "4.0"
graphql-core==3.2.3 ; python_version >= "3.10" and python_version < "4"
greenlet==1.1.3 ; python_version >= "3.10" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version < "4.0"
h11==0.12.0 ; python_version >= "3.10" and python_version < "4.0"
httpcore==0.15.0 ; python_version >= "3.10" and python_version < "4.0"
httptools==0.5.0 ; python_version >= "3.10" and python_version < "4.0"
httpx==0.23.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
imagesize==1.4.1 ; python_version >= "3.10" and python_version < "4.0"
iniconfig==1.1.1 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.1 ; python_version >= "3.10" and python_version < "4.0"
mypy-extensions==0.4.3 ; python_version >= "3.10" and python_version < "4.0"
none==0.1.0a2 ; python_version >= "3.10" and python_version < "4.0"
packaging==21.3 ; python_version >= "3.10" and python_version < "4.0"
passlib[bcrypt]==1.7.4 ; python_version >= "3.10" and python_version < "4.0"
password==0.2 ; python_version >= "3.10" and python_version < "4.0"
pathspec==0.10.1 ; python_version >= "3.10" and python_version < "4.0"
platformdirs==2.5.2 ; python_version >= "3.10" and python_version < "4.0"
pluggy==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
py==1.11.0 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.10" and python_version < "4.0"
pydantic==1.10.2 ; python_version >= "3.10" and python_version < "4.0"
pygments==2.13.0 ; python_version >= "3.10" and python_version < "4.0"
pyparsing==3.0.9 ; python_version >= "3.10" and python_version < "4.0"
pytest==7.1.3 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==0.21.0 ; python_version >= "3.10" and python_version < "4.0"
python-jose[cryptography]==3.3.0 ; python_version >= "3.10" and python_version < "4.0"
python-multipart==0.0.5 ; python_version >= "3.10" and python_version < "4.0"
pytz==2022.4 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
requests==2.28.1 ; python_version >= "3.10" and python_version < "4"
rfc3986[idna2008]==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
rfc3986[url]==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.10" and python_version < "4"
schema==0.7.5 ; python_version >= "3.10" and python_version < "4.0"
setuptools==65.4.1 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
snowballstemmer==2.2.0 ; python_version >= "3.10" and python_version < "4.0"
sphinx[docs]==3.5.4 ; python_version >= "3.10" and python_version < "4.0"
sphinxcontrib-applehelp==1.0.2 ; python_version >= "3.10" and python_version < "4.0"
sphinxcontrib-devhelp==1.0.2 ; python_version >= "3.10" and python_version < "4.0"
sphinxcontrib-htmlhelp==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
sphinxcontrib-jsmath==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
sphinxcontrib-qthelp==1.0.3 ; python_version >= "3.10" and python_version < "4.0"
sphinxcontrib-serializinghtml==1.1.5 ; python_version >= "3.10" and python_version < "4.0"
sphinxcontrib-websupport==1.2.4 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy[asyncio]==1.4.41 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.19.1 ; python_version >= "3.10" and python_version < "4.0"
strawberry-graphql[fastapi]==0.124.0 ; python_version >= "3.10" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.4.0 ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.12 ; python_version >= "3.10" and python_version < "4"
uvicorn[standard]==0.18.3 ; python_version >= "3.10" and python_version < "4.0"
uvloop==0.17.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.10" and python_version < "4.0"
watchfiles==0.17.0 ; python_version >= "3.10" and python_version < "4.0"
websockets==10.3 ; python_version >= "3.10" and python_version < "4.0"
